Refactor: place accounts' api-related code inside `api` subpackage
